+++ checks/codesign/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    openssl::ssl::SslConnector::builder(openssl::ssl::SslMethod::tls()).unwrap();
+}